TASK2. Properties file was added, Spring-config was changed to java/annotation, i18n was added
