+++ src/main/java/com/beshanov/config/AppConfig.java
@@ -1,0 +1,25 @@
+package com.beshanov.config;
+
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.*;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.context.support.ReloadableResourceBundleMessageSource;
+
+@Configuration
+@ComponentScan("com.beshanov")
+@PropertySource(value = "classpath:settings.properties")
+public class AppConfig {
+
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+    @Bean
+    public MessageSource messageSource() {
+        ReloadableResourceBundleMessageSource bundleMessageSource = new ReloadableResourceBundleMessageSource();
+        bundleMessageSource.setBasename("/i18n/labels");
+        bundleMessageSource.setDefaultEncoding("UTF-8");
+        return bundleMessageSource;
+    }
+}
