+++ src/main/java/com/beshanov/service/PresentationService.java
@@ -1,4 +1,4 @@
-package service;
+package com.beshanov.service;
 
 public interface PresentationService {
     void start();
