+++ src/main/resources/i18n/labels_en_US.properties
@@ -1,0 +1,2 @@
+type.your.name = Please, type your name
+result.for.user = Result for user {0} : {1} right answers from {2}
