+++ src/main/resources/i18n/labels_ru_RU.properties
@@ -1,0 +1,2 @@
+type.your.name = Пожалуйста, напишите ваше имя
+result.for.user = Результаты пользователя {0} : {1} правильных ответов из {2}
