+++ src/main/resources/questions_en_US.csv
@@ -1,0 +1,5 @@
+0,How many primitive types is in Java?,8
+1,How many bytes in int?,4
+2,String - mutable or immutable?,immutable
+3,How many reserved words in java?,50
+4,JVM JDK JRE: Which of the following contains javac?,JDK
