+++ src/main/resources/settings.properties
@@ -1,0 +1,2 @@
+questions.file=questions
+current.locale=en_US
