--- src/main/resources/spring_context.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-    <bean id="parser" class="utils.QuestionParser">
-        <constructor-arg name="fileName" value="questions.csv"></constructor-arg>
-    </bean>
-
-    <bean id="questionDao" class="dao.QuestionDaoImpl">
-        <constructor-arg name="questionParser" ref="parser"></constructor-arg>
-    </bean>
-
-    <bean id="questionService" class="service.QuestionServiceImpl">
-        <constructor-arg name="questionDao" ref="questionDao"></constructor-arg>
-    </bean>
-
-    <bean id="presentationService" class="service.PresentationServiceImpl">
-        <constructor-arg name="questionService" ref="questionService"></constructor-arg>
-    </bean>
-</beans>
